GUI: Fixed stuff.
Expanded 'GuiComponentDebug' to include more transformation tests.

'GuiButton' now accepts more colors.

A 'Logger' is required to create any 'GuiComponent'.

Constraints are applied every
 frame instead of on reshaping as transformations may change every frame
  without the shape changing. The resulting effect is quite satisfying.

Note that in 'GuiWindow', transforming the shape by resizing and
reposition is not perfect. This is hardly avoidable due to 'GuiComponent
.transformThis'.

'glScissor' is also fixed, however, it will be replaced as it does not
work for rotated components.

For mouse events, it seems like as long as at least one button is
pressed, the release event will not be fired for other buttons, and when
 all buttons are released, only one event is fired. The problem seems to
  reside in GLFW, so it is probably not fixable. A not very satisfying
  solution is implemented by firing 'IGuiEventListenerComponent
  .onMouseDragged' all at once for each button that is used to drag.

TODO:
- Replace 'glScissor'.
- Combine variables that are changeable externally into an object,
potentially allowing for the separation of model and view using the
state object for communication.
- Maybe use integers instead of double to ensure numerical stability.
Requires more testing to see whether this will impact the visual too
much.

Signed-off-by: William So <williamso26144225@gmail.com>