UI: Separate the renderer from the component
As a result, components window and button are separated into 2 classes
respectively, one of them is the general version, the other one is the
Minecraft version as the renderer is no longer tied to the component
implementation.  Other interfaces related to Minecraft have changed to
accommodate for this change.  Extra classes are created to help
implement a default renderer.

This change is mainly for paving the way to theming.  However, the
change should improve re-usability as well.

UI
- 'IUIComponent.initialize' and 'IUIComponent.removed' are added.  They
are basically the same as respectively the now-removed
'IUIComponentMinecraft.initialize' and 'IUIComponent.removed'.  Some
renaming may be required...
- Some packages related to theming has been created.  They are empty for
 now...

Miscellaneous
- Refactor code.
- Refactor stuff related to namespaces.

Status
Has been tested moderately.
Medium risk.

...

...

...

It turns out that 56a8ace73bc8968958c4798567b80f12f79084bd DID go wrong.
  The classes in 'components_test.in.xml' had not been refactored
  properly.  ;/
(4.9e-324% chance...  Inverse of XCOM chances?)

Signed-off-by: William So <williamso26144225@gmail.com>