UI: Refactor and fix 'IExtension' generic types
The 'IExtensionContainer' should not know the type of the extension as
this extension system works by the extension doing something on its own
instead of the container using the extension to do something.

Type safety can be ensured through the utility methods in
'IExtensionContainer' instead.

Also, fixed some messed up generics, such as '? super K' instead of '?
extends K' where 'K' is the key type accepted by 'IExtensionContainer'.
As the key is provided by the extension, the key provided by the
extension should extends 'k', not the other way around.

It's difficult to get generics right, isn't it?

Signed-off-by: William So <williamso26144225@gmail.com>