Improve XML unmarshalling by using JAXB
It is so much easier...  And any schema changes will produce a
compilation error instead of a runtime error, making changing schemas
much easier.

Old unmarshalling code are deleted.

However, the new system is not fully completed, we will need to make
some more changes to the component system, such as ensuring IDs are
unique, ensuring the correct behaviors of embedded
'IUIComponentManager's, etc.

This commit will fail to be built.  There will be compilation errors.
The build script is also not fully committed.  These will be resolved
within the next few commits.

Signed-off-by: William So <williamso26144225@gmail.com>