UI: Fix 'NoClassDefFoundError's due to 'LambdaMetafactory'
In order for it to work, the lookup object needs to have 'private'
access to the lookup class, and the lookup class and the target class
needs to be known by the same classloader.  Also, the lookup object
needs to be able to access the target symbol.  Thus, the calls to
'LambdaMetafactory' have been removed and replaced with direct calls to
the 'MethodHandle'.

Add some to-dos.

Signed-off-by: William So <williamso26144225@gmail.com>