[ui] Redesign 'IUIComponent' a bit
[ui] Redesign 'IUIComponent' a bit
- The design was a bit too parser-oriented.  This can be seen by the
existence of 'initializeRendererContainer'.  The parser design should
not affect the design of 'IUIComponent' and instead should only act as
an automated way to construct 'IUIComponent' within the constraint of
design of 'IUIComponent'.
- Removed 'initializeRendererContainer'.
- Move the name for the renderer to the construction arguments.
- Renderer name can be 'null' now.
- All component children handling except for adding and removing
children has been moved to the implementations of 'IUIViewComponent'.
This includes 'IHasBinding'.
- Update existing usage.

[ui] Create lifecycle system
- Create lifecycle events, which are 'bind', 'unbind', 'initialize', and
 'cleanup'.
- Create interfaces for the lifecycle system.
- 'bind' corresponds to 'unbind', and 'initialize' corresponds to
'cleanup'.
- The difference between the 'bind' set and the 'initialize' set is that
 'bind' set (un)freezes the structure of 'IUIInfrastructure' (note that
this does not freeze the hierarchy of 'IUIComponent'), while the
'initialize' set signals the use of 'IUIInfrastructure'.
- Retrofit existing similar design to use the lifecycle system.
- Update existing usage.

[ui] Flatten the rendering system
- Add method 'render' in 'IUIView' and 'IUIComponentRenderer'.
- This is likely to be improved in the near future with something akin
to 'IUIComponentPathResolver', which should allow for different ways of
rendering.  Originally the different ways of rendering will be
implemented by subclassing 'IUIComponentRenderer', however, this is now
removed here in favor of the above mentioned solution.  As the above
mentioned solution is not yet implemented so for now, rendering redraws
everything.
- Remove the Minecraft variants of 'IUIComponentRenderer'.
- Update existing usage.

[utilities] Create utilities, abstract classes
- Update existing usage.

Miscellaneous
- Refactor code.
- Improve code.
- Fix code.
- Can I stop repeating the above three messages?

Signed-off-by: William So <williamso26144225@gmail.com>